[REF] Minor cleanup and reformatting
