+++ delivery_shipit/__manifest__.py
@@ -1,7 +1,7 @@
 ##############################################################################
 #
-#    Author: Oy Tawasta OS Technologies Ltd.
-#    Copyright 2026 Oy Tawasta OS Technologies Ltd. (https://tawasta.fi)
+#    Author: Futural Oy.
+#    Copyright 2026 Futural Oy. (https://futural.fi)
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as
