+++ delivery_shipit/views/choose_delivery_carrier.xml
@@ -3,10 +3,15 @@
     <record id="view_choose_delivery_carrier_form_shipit" model="ir.ui.view">
         <field name="name">choose.delivery.carrier.form.shipit.auto.rate</field>
         <field name="model">choose.delivery.carrier</field>
-        <field name="inherit_id" ref="stock_delivery.choose_delivery_carrier_view_form" />
+        <field
+            name="inherit_id"
+            ref="stock_delivery.choose_delivery_carrier_view_form"
+        />
         <field name="arch" type="xml">
             <xpath expr="//button[@name='update_price']" position="attributes">
-                <attribute name="invisible">delivery_type in ('fixed', 'base_on_rule', 'shipit')</attribute>
+                <attribute
+                    name="invisible"
+                >delivery_type in ('fixed', 'base_on_rule', 'shipit')</attribute>
             </xpath>
         </field>
     </record>
