+++ delivery_shipit/wizards/__init__.py
@@ -1,2 +1,3 @@
-from . import shipit_pickup_point_wizard
 from . import choose_delivery_carrier
+from . import shipit_pickup_point_wizard
+from . import shipit_pickup_point_wizard_line
